feat: login PIN, tarjetas, mermas con aprobacion, ventas con dropdown

+++ app/src/main/java/org/luisito/gestor360/data/models/CartItem.kt
@@ -1,7 +1,8 @@
 package org.luisito.gestor360.data.models
 
+/** Ítem del carrito de venta. Vive solo en memoria mientras se arma la venta. */
 data class CartItem(
-    val productId: Int,
+    val productId: Long,
     val nombre: String,
     val precio: Double,
     val cantidad: Double,
